Client/search feat

+++ src/components/Main/NodeOption.jsx
@@ -3,7 +3,7 @@ import React from "react"
 import { useSelector, useDispatch } from 'react-redux';
 import { closeNodeoptionModal, closeNodesettingModal } from '../../redux/modalstatus'
 import { addNodeData } from '../../redux/nodemap';
-// import { addEdgeData } from '../../redux/node';
+
 import axios from 'axios';
 import dotenv from 'dotenv';
 dotenv.config();
@@ -16,6 +16,7 @@ const NodeOption = ()=>{
     const [color, setColor] = useState("");
     const { parentNode } = useSelector(state => state.node);
     const { nodeData } = useSelector(state => state.node);
+
     // const { edgeData } = useSelector(state => state.node);
     const state = useSelector(state=>state)
     console.log(state)
@@ -23,34 +24,52 @@ const NodeOption = ()=>{
     const closeModal = () => {
         dispatch(closeNodeoptionModal());
     }
+    const checkNameValidation = (target) => {
+        let find = nodeData.filter((node) => {
+            return node.nodeName === target;
+        });
+        if (find.length !== 0) {
+            return true;
+        }
+        return false;
+    }
 
     const submitHandler = () => {
         const token = localStorage.getItem('token');
-        dispatch(closeNodesettingModal());
-        axios.post(`${SERVER_API}/nodemap/node`,
-            {
-                nodeColor: color,
-                nodeName: nodeName.current.value,
-                target: nodeData.length,
-                source: parentNode.id
-            },
-            { headers: { withCredentials: true, 'authorization': `Bearer ${token}` } })
-            .then(res => res.data)
-            .then(data => {
-                dispatch(closeNodeoptionModal());
-                dispatch(addNodeData({ ...data.nodeData, id: nodeData.length },
-                    { ...data.edgeData }));
-                return data;
-            })
-            .catch(error => {  //TODO:TOKEN확인여부를 가지고 로그인 창을 띄움
-                if (error.status === 400) {
-                    //! 세션만료 모달, 로그인 해제
-                    localStorage.clear();
-                   //TODO:로그아웃 해야함
-                    window.location.reload();
-                    return alert('세션이 만료되었습니다. 다시 로그인 해주세요');
-                }
-            });
+        if (checkNameValidation(nodeName.current.value)) {
+            return alert("이미 존재하는 노드입니다.");
+        }
+        if (!nodeName || !color) {
+            return alert("옵션 정보를 모두 입력해주세요.");
+        } else {
+            console.log('등록완료!');
+            dispatch(closeNodesettingModal());
+            axios.post(`${SERVER_API}/nodemap/node`,
+                {
+                    nodeColor: color,
+                    nodeName: nodeName.current.value,
+                    target: nodeName.current.value,
+                    source: parentNode.nodeName
+                },
+                { headers: { withCredentials: true, 'authorization': `Bearer ${token}` } })
+                .then(res => res.data)
+                .then(data => {
+                    dispatch(closeNodeoptionModal());
+                    
+                    dispatch(addNodeData({ id: data.nodeData.nodeName, ...data.nodeData },
+                        { ...data.edgeData }));
+                    return data;
+                })
+                .catch(error => {  //TODO:TOKEN확인여부를 가지고 로그인 창을 띄움
+                    if (error.status === 400) {
+                        //! 세션만료 모달, 로그인 해제
+                        localStorage.clear();
+                        //TODO:로그아웃 해야함
+                        window.location.reload();
+                        return alert('세션이 만료되었습니다. 다시 로그인 해주세요');
+                    }
+                });
+        }
     }
     return (
         <div className='nodeoption-darkbackground'>
